Add tests for result transformation
