+++ src/main/kotlin/apoc/nlp/aws/AWSProcedures.kt
@@ -14,14 +14,6 @@ class AWSProcedures {
     @JvmField
     var log: Log? = null
 
-    //    @Procedure(value = "apoc.ai.aws.sentiment", mode = Mode.READ)
-//    @Description("Provides a sentiment analysis for provided text")
-//    fun sentiment(@Name("apiKey") apiKey: String,
-//                  @Name("apiSecret") apiSecret: String,
-//                  @Name("data") data: Any,
-//                  @Name(value = "config", defaultValue = "{}") config: Map<String, Any>)
-//            : Stream<AIMapResult> = AWSClient(apiKey, apiSecret, log!!).sentiment(data, config).stream()
-//
     @Procedure(value = "apoc.nlp.aws.entities.stream", mode = Mode.READ)
     @Description("Returns a stream of entities for provided text")
     fun entities(@Name("source") source: Any,
@@ -30,27 +22,15 @@ class AWSProcedures {
         verifySource(source)
         val nodeProperty = getNodeProperty(config)
         verifyNodeProperty(source, nodeProperty)
-
         verifyKey(config, "key")
         verifyKey(config, "secret")
 
         val client = AWSClient(config, log!!)
-
         val detectEntitiesResult = client.entities(source, config)
         val resultList = detectEntitiesResult!!.resultList
         val errorList = detectEntitiesResult.errorList
 
-        return convert(source).mapIndexed { index, node  -> transform(index, node, resultList, errorList) }.stream()
-    }
-
-    private fun transform(index: Int, node: Node, resultList: List<BatchDetectEntitiesItemResult>, errorList: List<BatchItemError>) : NodeWithMapResult {
-        val result = resultList.find { result -> result.index == index }
-        return if (result != null) {
-            NodeWithMapResult.withResult(node, JsonUtil.OBJECT_MAPPER!!.convertValue(result, Map::class.java) as Map<String, Any?>)
-        } else {
-            val err = errorList.find { error -> error.index == index }
-            NodeWithMapResult.withError(node, mapOf("code" to err?.errorCode, "message" to err?.errorMessage))
-        }
+        return convert(source).mapIndexed { index, node  -> Companion.transformResults(index, node, resultList, errorList) }.stream()
     }
 
     private fun convert(source: Any) : List<Node> {
@@ -94,6 +74,15 @@ class AWSProcedures {
             throw IllegalArgumentException("$node does not have property `$nodeProperty`. Property can be configured using parameter `nodeProperty`.")
         }
     }
+
+//    @Procedure(value = "apoc.ai.aws.sentiment", mode = Mode.READ)
+//    @Description("Provides a sentiment analysis for provided text")
+//    fun sentiment(@Name("apiKey") apiKey: String,
+//                  @Name("apiSecret") apiSecret: String,
+//                  @Name("data") data: Any,
+//                  @Name(value = "config", defaultValue = "{}") config: Map<String, Any>)
+//            : Stream<AIMapResult> = AWSClient(apiKey, apiSecret, log!!).sentiment(data, config).stream()
+//
 //
 //    @Procedure(value = "apoc.ai.aws.keyPhrases", mode = Mode.READ)
 //    @Description("Provides a entity analysis for provided text")
@@ -109,5 +98,16 @@ class AWSProcedures {
 //               @Name("apiSecret") apiSecret: String,
 //               @Name("data") data: Any,
 //               @Name(value = "config", defaultValue = "{}") config: Map<String, Any>): Stream<AIMapResult> = Stream.empty()
+    companion object {
+        fun transformResults(index: Int, node: Node, resultList: List<BatchDetectEntitiesItemResult>, errorList: List<BatchItemError>) : NodeWithMapResult {
+            val result = resultList.find { result -> result.index == index }
+            return if (result != null) {
+                NodeWithMapResult.withResult(node, JsonUtil.OBJECT_MAPPER!!.convertValue(result, Map::class.java) as Map<String, Any?>)
+            } else {
+                val err = errorList.find { error -> error.index == index }
+                NodeWithMapResult.withError(node, mapOf("code" to err?.errorCode, "message" to err?.errorMessage))
+            }
+        }
+    }
 
 }
