+++ src/test/kotlin/apoc/nlp/aws/AWSProceduresErrorsTest.kt
@@ -91,5 +91,4 @@ class AWSProceduresErrorsTest {
         }
         assertThat(exception.message, containsString("java.lang.IllegalArgumentException: Missing parameter `secret`"))
     }
-
 }
